+++ runway/data_types.py
@@ -14,8 +14,42 @@ from PIL import Image
 from .utils import is_url, extract_tarball, try_cast_np_scalar, download_file, random_color_map
 from .exceptions import MissingArgumentError, InvalidArgumentError
 
+class BaseType(object):
+    """An abstract class that defines a base data type interface. This type
+    should be used as the base class of new data types, never directly.
 
-class any(object):
+    :param data_type: The data type represented as a string
+    :type data_type: string
+    :param description: A description of this variable and how its used in the model,
+        defaults to None
+    :type description: string, optional
+    """
+
+    def __init__(self, data_type, description=None):
+        self.type = data_type
+        self.description = description
+
+        # The name property is assigned after contruction through direct
+        # property assignment (e.g. `txt = text(); txt.name = 'some_name' `)
+        # It is the responsibility of the RunwayModel's setup() and command()
+        # functions to assign names to runway.data_types based on the dictionary
+        # keys
+        self.name = None
+
+    def serialize(self, value):
+        raise NotImplementedError()
+
+    def deserialize(self, value):
+        raise NotImplementedError()
+
+    def to_dict(self):
+        return {
+            'name': self.name,
+            'type': self.type,
+            'description': self.description
+        }
+
+class any(BaseType):
     """A generic data type. The value this data type takes must be serializable to JSON.
 
     .. code-block:: python
@@ -26,19 +60,20 @@ class any(object):
         from your_code import model
 
         # an example of passing your own yaml configuration using an "any" data_type and PyYAML
-        @runway.setup(options={ "configuration": any(name="yaml") })
+        @runway.setup(options={ "configuration": any() })
         def setup(opts):
             # parse the configuration string as yaml, and then pass the resulting
             # object as the configuration to your model
             config = yaml.load(opts["configuration"])
             return model(config)
 
-    :param name: The name associated with this variable, defaults to "field"
-    :type name: string, optional
+    :param description: A description of this variable and how its used in the model,
+        defaults to None
+    :type description: string, optional
     """
 
-    def __init__(self, name=None):
-        self.name = name
+    def __init__(self, description=None):
+        super(any, self).__init__('any', description=description)
 
     def serialize(self, v):
         return v
@@ -47,12 +82,9 @@ class any(object):
         return v
 
     def to_dict(self):
-        ret = {}
-        ret['type'] = 'any'
-        ret['name'] = self.name or 'field'
-        return ret
+        return super(any, self).to_dict()
 
-class array(object):
+class array(BaseType):
     """A data type representing an array (list) of other runway.data_type objects.
 
     .. code-block:: python
@@ -67,22 +99,23 @@ class array(object):
 
     :param item_type: A runway.data_type class, or an instance of a runway.data_type class
     :type item_type: runway.data_type
-    :param name: The name associated with this variable, defaults to "{item_type}_array" \
-        (e.g. "image_array")
-    :type name: string, optional
+    :param description: A description of this variable and how its used in the model,
+        defaults to None
+    :type description: string, optional
     :param min_length: The minimum number of elements required to be in the array, defaults to 0
     :type min_length: int, optional
     :param max_length: The maximum number of elements allowed to be in the array, defaults to None
     :type max_length: int, optional
     :raises MissingArgumentError: A missing argument error if item_type is not specified
     """
-    def __init__(self, item_type=None, name=None, min_length=0, max_length=None):
+    def __init__(self, item_type=None, description=None, min_length=0, max_length=None):
+        super(array, self).__init__('array', description=description)
         if item_type is None: raise MissingArgumentError('item_type')
         if inspect.isclass(item_type):
             self.item_type = item_type()
         else:
             self.item_type = item_type
-        self.name = name or '%s_array' % self.item_type.name
+        self.item_type.name = '%s_array_item' % self.item_type.type
         self.min_length = min_length
         self.max_length = max_length
 
@@ -93,16 +126,14 @@ class array(object):
         return [self.item_type.serialize(item) for item in items]
 
     def to_dict(self):
-        ret = {}
-        ret['name'] = self.name
-        ret['type'] = 'array'
+        ret = super(array, self).to_dict()
         ret['itemType'] = self.item_type.to_dict()
         ret['minLength'] = self.min_length
         if self.max_length: ret['maxLength'] = self.max_length
         return ret
 
 
-class image(object):
+class image(BaseType):
     """A data type representing an image. Images represent PIL or numpy
     images but are passed to and from the Model SDK as base64 encoded data URI
     strings over the network
@@ -128,8 +159,9 @@ class image(object):
             # encoded data URI string by the @runway.command() decorator.
             return { "image": img }
 
-    :param name: The name associated with this variable, defaults to "image"
-    :type name: string, optional
+    :param description: A description of this variable and how its used in the model,
+        defaults to None
+    :type description: string, optional
     :param channels: The number of channels in the image, defaults to 3. \
         E.g. an "rgb" image has 3 channels while an "rgba" image has 4.
     :type channels: int, optional
@@ -146,8 +178,8 @@ class image(object):
     :param height: The height of the image, defaults to None
     :type height: int, optional
     """
-    def __init__(self, name=None, channels=3, min_width=None, min_height=None, max_width=None, max_height=None, width=None, height=None):
-        self.name = name or 'image'
+    def __init__(self, description=None, channels=3, min_width=None, min_height=None, max_width=None, max_height=None, width=None, height=None):
+        super(image, self).__init__('image', description=description)
         self.channels = channels
         self.min_width = min_width
         self.min_height = min_height
@@ -168,15 +200,13 @@ class image(object):
         elif issubclass(type(value), Image.Image):
             im_pil = value
         else:
-            raise InvalidArgumentError('value is not a PIL or numpy image')
+            raise InvalidArgumentError(self.name or self.type, 'value is not a PIL or numpy image')
         buffer = IO()
         im_pil.save(buffer, format='JPEG')
         return 'data:image/jpeg;base64,' + base64.b64encode(buffer.getvalue()).decode('utf8')
 
     def to_dict(self):
-        ret = {}
-        ret['type'] = 'image'
-        ret['name'] = self.name
+        ret = super(image, self).to_dict()
         ret['channels'] = self.channels
         if self.min_width: ret['minWidth'] = self.min_width
         if self.max_width: ret['maxWidth'] = self.max_width
@@ -187,7 +217,7 @@ class image(object):
         return ret
 
 
-class vector(object):
+class vector(BaseType):
     """A data type representing a vector of floats.
 
     .. code-block:: python
@@ -204,26 +234,27 @@ class vector(object):
             rand = np.random.random_sample(args["length"])
             return { "vector": vec.deserialize(rand) }
 
+    :param description: A description of this variable and how its used in the model,
+        defaults to None
+    :type description: string, optional
     :param length: The number of elements in the vector
-    :type length: int
-    :param name: The name associated with this variable, defaults to "vector"
-    :type name: string, optional
+    :type length: int, inferred if a default vector is specified
     :param sampling_mean: The mean of the sample the vector is drawn from, defaults to 0
     :type sampling_mean: float, optional
     :param sampling_std: The standard deviation of the sample the vector is drawn from, defaults to 1
     :type sampling_std: float, optional
     :raises MissingArgumentError: A missing argument error if length is not specified
     """
-    def __init__(self, length=None, name=None, default=None, sampling_mean=0, sampling_std=1):
+    def __init__(self, description=None, length=None, default=None, sampling_mean=0, sampling_std=1):
+        super(vector, self).__init__('vector', description=description)
         if default is not None:
             if length is None:
                 length = len(default)
             elif len(default) != length:
                 msg = 'default argument does not match expected length'
-                raise InvalidArgumentError(msg)
+                raise InvalidArgumentError(self.name or self.type, msg)
         if length is None:
             raise MissingArgumentError('length')
-        self.name = name or 'vector'
         self.length = length
         self.sampling_mean = sampling_mean
         self.sampling_std = sampling_std
@@ -236,9 +267,7 @@ class vector(object):
         return value.tolist()
 
     def to_dict(self):
-        ret = {}
-        ret['type'] = 'vector'
-        ret['name'] = self.name
+        ret = super(vector, self).to_dict()
         ret['length'] = self.length
         ret['samplingMean'] = self.sampling_mean
         ret['samplingStd'] = self.sampling_std
@@ -246,7 +275,7 @@ class vector(object):
         return ret
 
 
-class category(object):
+class category(BaseType):
     """A categorical data type that allows you to specify a variable's value \
         as a member of a set list of choices.
 
@@ -262,8 +291,9 @@ class category(object):
             def setup(opts):
                 print("The selected pixel order is {}".format(opts["pixel_order"]))
 
-        :param name: The name associated with this variable, defaults to "category"
-        :type name: string, optional
+        :param description: A description of this variable and how its used in the model,
+            defaults to None
+        :type description: string, optional
         :param choices: A list of categories, defaults to None
         :type choices: A list of strings
         :param default: A default list of categories, defaults to None
@@ -275,34 +305,32 @@ class category(object):
             choices list.
     """
 
-    def __init__(self, name=None, choices=None, default=None):
+    def __init__(self, description=None, choices=None, default=None):
+        super(category, self).__init__('category', description=description)
         if choices is None or len(choices) == 0: raise MissingArgumentError('choices')
         if default is not None and default not in choices:
             msg = 'default argument {} is not in choices list'.format(default)
-            raise InvalidArgumentError(msg)
-        self.name = name or 'category'
+            raise InvalidArgumentError(self.name or self.type, msg)
         self.choices = choices
         self.default = default or self.choices[0]
 
     def deserialize(self, value):
         if value not in self.choices:
             msg = 'category value "%s" does not appear in choices list.' % value
-            raise InvalidArgumentError(self.name, msg)
+            raise InvalidArgumentError(self.name or self.type, msg)
         return value
 
     def serialize(self, value):
         return value
 
     def to_dict(self):
-        ret = {}
-        ret['type'] = 'category'
-        ret['name'] = self.name
+        ret = super(category, self).to_dict()
         ret['oneOf'] = self.choices
         ret['default'] = self.default
         return ret
 
 
-class number(object):
+class number(BaseType):
     """A basic number data type. Instantiate this class to create a new runway model variable.
 
     .. code-block:: python
@@ -314,8 +342,9 @@ class number(object):
         def setup(opts):
             print("The number of samples is {}".format(opts["number_of_samples"]))
 
-    :param name: The name associated with this variable, defaults to None
-    :type name: string, optional
+    :param description: A description of this variable and how its used in the model,
+        defaults to None
+    :type description: string, optional
     :param default: A default value for this number variable, defaults to 0
     :type default: float, optional
     :param min: The minimum allowed value of this number type, defaults to 0
@@ -328,8 +357,8 @@ class number(object):
     :type step: float, optional
     """
 
-    def __init__(self, name=None, default=0, min=0, max=1, step=1):
-        self.name = name or 'number'
+    def __init__(self, description=None, default=0, min=0, max=1, step=1):
+        super(number, self).__init__('number', description=description)
         self.default = default
         self.min = min
         self.max = max
@@ -342,9 +371,7 @@ class number(object):
         return try_cast_np_scalar(value)
 
     def to_dict(self):
-        ret = {}
-        ret['type'] = 'number'
-        ret['name'] = self.name
+        ret = super(number, self).to_dict()
         ret['default'] = self.default
         ret['min'] = self.min
         ret['max'] = self.max
@@ -352,7 +379,7 @@ class number(object):
         return ret
 
 
-class text(object):
+class text(BaseType):
     """A basic text data type. Used to represent strings. \
         Instantiate this class to create a new runway model variable.
 
@@ -365,18 +392,19 @@ class text(object):
         def setup(opts):
             print("The selected flavor is {}".format(opts["flavor"]))
 
-    :param name: The name associated with this variable, defaults to "text"
-    :type name: string, optional
+    :param description: A description of this variable and how its used in the model,
+        defaults to None
+    :type description: string, optional
     :param default: The default value for this text variable, defaults to ''
-    :type default: str, optional
+    :type default: string, optional
     :param min_length: The minimum character length of this text variable, defaults to 0
     :type min_length: int, optional
     :param max_length: The maximum character length of this text variable, \
         defaults to None, which allows text to be of any maximum length
     :type max_length: int, optional
     """
-    def __init__(self, name=None, default='', min_length=0, max_length=None):
-        self.name = name or 'text'
+    def __init__(self, description=None, default='', min_length=0, max_length=None):
+        super(text, self).__init__('text', description=description)
         self.default = default
         self.min_length = min_length
         self.max_length = max_length
@@ -388,16 +416,14 @@ class text(object):
         return str(value)
 
     def to_dict(self):
-        ret = {}
-        ret['type'] = 'text'
-        ret['name'] = self.name
+        ret = super(text, self).to_dict()
         ret['default'] = self.default
         ret['minLength'] = self.min_length
         if self.max_length: ret['maxLength'] = self.max_length
         return ret
 
 
-class file(object):
+class file(BaseType):
     """A data type that represents a file or directory. The file can be a local \
         resource on disk or a remote resource loaded over HTTP. \
         Instantiate this class to create a new runway model variable.
@@ -414,16 +440,17 @@ class file(object):
             result = do_something_with(args["directory"])
             return { "result": "success" if result else "failure" }
 
-    :param name: The name of this variable, defaults to None
-    :type name: string, optional
+    :param description: A description of this variable and how its used in the model,
+        defaults to None
+    :type description: string, optional
     :param is_directory: Does this variable represent a directory instead of a file? Defaults to False.
     :type is_directory: bool, optional
     :param extension: Accept only files of this extension.
     :type extension: string, optional
     """
 
-    def __init__(self, name=None, is_directory=False, extension=None):
-        self.name = name or 'file'
+    def __init__(self, description=None, is_directory=False, extension=None):
+        super(file, self).__init__('file', description=description)
         self.is_directory = is_directory
         self.extension = extension
 
@@ -436,18 +463,16 @@ class file(object):
                 return downloaded_path
         else:
             if not os.path.exists(path_or_url):
-                raise InvalidArgumentError('file path provided does not exist')
+                raise InvalidArgumentError(self.name or self.type, 'file path provided does not exist')
             if self.extension and not path_or_url.endswith(self.extension):
-                raise InvalidArgumentError('file path does not have expected extension')
+                raise InvalidArgumentError(self.name or self.type, 'file path does not have expected extension')
             return path_or_url
 
     def serialize(self, value):
         return value
 
     def to_dict(self):
-        ret = {}
-        ret['type'] = 'file'
-        ret['name'] = self.name
+        ret = super(file, self).to_dict()
         if self.is_directory: ret['isDirectory'] = self.is_directory
         if self.extension: ret['extension'] = self.extension
         return ret