+++ tests/test_data_types.py
@@ -31,6 +31,9 @@ def check_expected_contents_for_0057_tar_download(path):
         assert f.read() == '# Runway Python SDK\n'
 
 # BASIC TESTS FOR ALL DATA TYPES -----------------------------------------------
+def test_data_type_interface_base_type():
+    check_data_type_interface(BaseType)
+
 def test_data_type_interface_any():
     check_data_type_interface(any)
 
@@ -55,16 +58,64 @@ def test_data_type_interface_text():
 def test_data_type_interface_file():
     check_data_type_interface(file)
 
+# BASE TYPE --------------------------------------------------------------------
+def test_base_type_to_dict():
+
+    base_type = BaseType('base', description='Some description.')
+    obj = base_type.to_dict()
+    assert obj['type'] == 'base'
+    assert obj['description'] == 'Some description.'
+
+# The BaseType is an abstract class that requires its serialize/deserialize
+# methods to be overwritten by subclasses
+def test_base_type_serialize_not_implemented():
+    base_type = BaseType('base')
+    with pytest.raises(NotImplementedError):
+        base_type.serialize('test')
+
+def test_base_type_deserialize_not_implemented():
+    base_type = BaseType('base')
+    with pytest.raises(NotImplementedError):
+        base_type.deserialize('test')
+
+# ANY --------------------------------------------------------------------------
+def test_any_to_dict():
+    a = any()
+    obj = a.to_dict()
+    assert obj['type'] == 'any'
+    assert obj['description'] == None
+
+def test_any_serialization():
+    a = any()
+    assert a.serialize(512) == 512
+    assert a.serialize(512.5) == 512.5
+    assert a.serialize('512') == '512'
+    assert a.serialize(None) == None
+    assert a.serialize(True) == True
+    assert a.serialize([]) == []
+    assert a.serialize({}) == {}
+
+def test_any_deserialize():
+    a = any()
+    assert a.deserialize(512) == 512
+    assert a.deserialize(512.5) == 512.5
+    assert a.deserialize('512') == '512'
+    assert a.deserialize(None) == None
+    assert a.deserialize(True) == True
+    assert a.deserialize([]) == []
+    assert a.deserialize({}) == {}
+
 # TEXT -------------------------------------------------------------------------
 def test_text_to_dict():
     default = 'Some default text'
-    txt = text(default=default, min_length=1, max_length=20)
+    description = 'A description about this variable.'
+    txt = text(default=default, description=description, min_length=1, max_length=20)
     obj = txt.to_dict()
     assert obj['type'] == 'text'
-    assert obj['name'] == 'text'
     assert obj['default'] == default
     assert obj['minLength'] == 1
     assert obj['maxLength'] == 20
+    assert obj['description'] == description
 
 def test_text_serialization():
     txt = text()
@@ -77,14 +128,15 @@ def test_text_deserialize():
 # NUMBER -----------------------------------------------------------------------
 def test_number_to_dict():
     default = 42
-    num = number(default=default, min=10, max=100)
+    description = 'A description about this variable.'
+    num = number(default=default, description=description, min=10, max=100)
     obj = num.to_dict()
     assert obj['type'] == 'number'
-    assert obj['name'] == 'number'
     assert obj['default'] == default
     assert obj['min'] == 10
     assert obj['max'] == 100
     assert obj['step'] == 1
+    assert obj['description'] == description
 
 def test_number_serialization():
     assert 1 == number().serialize(1)
@@ -102,13 +154,18 @@ def test_number_serialize_numpy_scalar():
 
 # ARRAY ------------------------------------------------------------------------
 def test_array_to_dict():
-    arr = array(item_type=text, min_length=5, max_length=10)
+    description = 'A description about this variable.'
+    arr = array(item_type=text, description=description, min_length=5, max_length=10)
     obj = arr.to_dict()
-    assert obj['name'] == 'text_array'
+
     assert obj['type'] == 'array'
-    assert obj['itemType'] == text().to_dict()
     assert obj['minLength'] == 5
     assert obj['maxLength'] == 10
+    assert obj['description'] == description
+
+    tmp = text()
+    tmp.name = 'text_array_item'
+    assert obj['itemType'] == tmp.to_dict()
 
 def test_array_no_item_type():
     with pytest.raises(MissingArgumentError):
@@ -135,13 +192,14 @@ def test_array_deserialization():
 
 # VECTOR -----------------------------------------------------------------------
 def test_vector_to_dict():
-    vec = vector(length=128, sampling_mean=0, sampling_std=1)
+    description = 'A description about this variable.'
+    vec = vector(length=128, description=description, sampling_mean=0, sampling_std=1)
     obj = vec.to_dict()
-    assert obj['name'] == 'vector'
     assert obj['type'] == 'vector'
     assert obj['length'] == 128
     assert obj['samplingMean'] == 0
     assert obj['samplingStd'] == 1
+    assert obj['description'] == description
 
 def test_vector_no_item_type():
     with pytest.raises(MissingArgumentError):
@@ -173,12 +231,13 @@ def test_vector_default_no_length_arg():
 
 # CATEGORY ---------------------------------------------------------------------
 def test_category_to_dict():
-    cat = category(choices=['one', 'two', 'three'], default='two')
+    description = 'A description about this variable.'
+    cat = category(choices=['one', 'two', 'three'], default='two', description=description)
     obj = cat.to_dict()
-    assert obj['name'] == 'category'
     assert obj['type'] == 'category'
     assert obj['oneOf'] == ['one', 'two', 'three']
     assert obj['default'] == 'two'
+    assert obj['description'] == description
 
 def test_category_serialization():
     cat = category(choices=['one', 'two', 'three'], default='two')
@@ -217,15 +276,16 @@ def test_category_deserialized_value_is_not_in_choices():
 def test_file_to_dict():
     f = file()
     obj = f.to_dict()
-    assert obj['name'] == 'file'
     assert obj['type'] == 'file'
+    assert obj['description'] == None
 
 def test_file_to_dict_directory():
-    f = file(is_directory=True)
+    description = 'A description about this variable.'
+    f = file(is_directory=True, description=description)
     obj = f.to_dict()
-    assert obj['name'] == 'file'
     assert obj['type'] == 'file'
     assert obj['isDirectory'] == True
+    assert obj['description'] == description
 
 def test_file_serialization_base():
     f = file()
@@ -313,12 +373,12 @@ def test_image_to_dict():
     img = image(channels=3, min_width=128, min_height=128, max_width=512, max_height=512)
     obj = img.to_dict()
     assert obj['type'] == 'image'
-    assert obj['name'] == 'image'
     assert obj['channels'] == 3
     assert obj['minWidth'] == 128
     assert obj['maxWidth'] == 512
     assert obj['minHeight'] == 128
     assert obj['maxHeight'] == 512
+    assert obj['description'] == None
 
 def test_image_serialize_and_deserialize():
     directory = os.path.dirname(os.path.realpath(__file__))